change some es5 to es6
