+++ client2/src/components/common/navheader.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import { Navbar } from 'react-bootstrap';
+
+export default (props) => {
+	return (
+		<Navbar.Header>
+        	<Navbar.Brand>
+        		{props.title}
+          </Navbar.Brand>
+    </Navbar.Header>
+		)
+}
