+++ package.json
@@ -36,6 +36,7 @@
     "imports-loader": "^0.6.4",
     "jest-cli": "^0.8.2",
     "localStorage": "^1.0.3",
+    "localeval": "^15.2.3",
     "lodash": "^3.10.1",
     "method-override": "~2.1.2",
     "morgan": "~1.2.2",
